Use pathping instead of powershell Find-NetRoute -RemoteIPAddress.
The number of hops can be controlled when using `pathping` and starting
a powershell is sometimes a very slow process.
