+++ integration_test/cloud_config_test.go
@@ -52,7 +52,14 @@ func (s *notConnectedSuite) Test_CloudNeverProxy() {
 		timeout *= 5
 	}
 	s.Eventuallyf(func() bool {
-		defer itest.TelepresenceDisconnectOk(ctx)
+		defer func() {
+			stdout, stderr, err := itest.Telepresence(ctx, "quit")
+			dlog.Infof(ctx, "stdout: %q", stdout)
+			dlog.Infof(ctx, "stderr: %q", stderr)
+			if err != nil {
+				dlog.Error(ctx, err)
+			}
+		}()
 		stdout, stderr, err := itest.Telepresence(ctx, "connect", "--namespace", s.AppNamespace(), "--manager-namespace", s.ManagerNamespace())
 		dlog.Infof(ctx, "stdout: %q", stdout)
 		dlog.Infof(ctx, "stderr: %q", stderr)
@@ -92,6 +99,7 @@ func (s *notConnectedSuite) Test_CloudNeverProxy() {
 			return false
 		}
 
+		dlog.Infof(ctx, "Success! Never-proxied IP %s is not reachable", ip)
 		return true
 	}, timeout, 5*time.Second, "never-proxy not updated in %s", timeout)
 }
@@ -143,7 +151,7 @@ func (s *notConnectedSuite) Test_RootdCloudLogLevel() {
 			currentLine++
 		}
 		return levelSet
-	}, 20*time.Second, 5*time.Second, "Root log level not updated in 20 seconds")
+	}, 60*time.Second, 5*time.Second, "Root log level not updated in 20 seconds")
 
 	// Make sure the log level was set back after disconnect
 	rootLog, err = os.Open(rootLogName)
@@ -229,7 +237,7 @@ func (s *notConnectedSuite) Test_UserdCloudLogLevel() {
 			currentLine++
 		}
 		return levelSet
-	}, 20*time.Second, 5*time.Second, "Connector log level not updated in 20 seconds")
+	}, 60*time.Second, 5*time.Second, "Connector log level not updated in 20 seconds")
 
 	// Make sure the log level was set back after disconnect
 	logF, err = os.Open(logName)
