+++ pkg/routing/routing_linux.go
@@ -136,7 +136,7 @@ msgLoop:
 	return routes, nil
 }
 
-func getRoute(ctx context.Context, routedNet *net.IPNet) (*Route, error) {
+func getOsRoute(ctx context.Context, routedNet *net.IPNet) (*Route, error) {
 	ip := routedNet.IP
 	cmd := dexec.CommandContext(ctx, "ip", "route", "get", ip.String())
 	cmd.DisableLogging = true
